// src/database/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
// ...
      payments: {
        Row: {
          amount: number
          created_at: string
          created_by: string | null
          description: string | null
          id: number
          payment_date: string | null
          payment_method: Database["public"]["Enums"]["payment_method_type"]
          reference_no: string | null
          repair_order_id: string
          store_id: string | null
          udpated_at: string | null
          updated_by: string | null
        }
        Insert: {
          amount: number
          created_at?: string
          created_by?: string | null
          description?: string | null
          id?: number
          payment_date?: string | null
          payment_method: Database["public"]["Enums"]["payment_method_type"]
          reference_no?: string | null
          repair_order_id: string
          store_id?: string | null
          udpated_at?: string | null
          updated_by?: string | null
        }
        Update: {
          amount?: number
          created_at?: string
          created_by?: string | null
          description?: string | null
          id?: number
          payment_date?: string | null
          payment_method?: Database["public"]["Enums"]["payment_method_type"]
          reference_no?: string | null
          repair_order_id?: string
          store_id?: string | null
          udpated_at?: string | null
          updated_by?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "payments_created_by_fkey"
            columns: ["created_by"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "payments_repair_order_id_fkey"
            columns: ["repair_order_id"]
            isOneToOne: false
            referencedRelation: "repair_order"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "payments_repair_order_id_fkey"
            columns: ["repair_order_id"]
            isOneToOne: false
            referencedRelation: "repair_orders"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "payments_store_id_fkey"
            columns: ["store_id"]
            isOneToOne: false
            referencedRelation: "store"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "payments_store_id_fkey"
            columns: ["store_id"]
            isOneToOne: false
            referencedRelation: "store_services"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "payments_updated_by_fkey"
            columns: ["updated_by"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      repair_order_line: {
        Row: {
          created_at: string
          id: number
          name: string | null
          price: number | null
          quantity: number
          repair_order_id: string | null
          service_id: number | null
          store_id: string
          updated_at: string | null
        }
        Insert: {
          created_at?: string
          id?: number
          name?: string | null
          price?: number | null
          quantity?: number
          repair_order_id?: string | null
          service_id?: number | null
          store_id: string
          updated_at?: string | null
        }
        Update: {
          created_at?: string
          id?: number
          name?: string | null
          price?: number | null
          quantity?: number
          repair_order_id?: string | null
          service_id?: number | null
          store_id?: string
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "repair_order_line_repair_order_id_fkey"
            columns: ["repair_order_id"]
            isOneToOne: false
            referencedRelation: "repair_order"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "repair_order_line_repair_order_id_fkey"
            columns: ["repair_order_id"]
            isOneToOne: false
            referencedRelation: "repair_orders"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "repair_order_line_service_id_fkey"
            columns: ["service_id"]
            isOneToOne: false
            referencedRelation: "service"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "repair_order_line_service_id_fkey"
            columns: ["service_id"]
            isOneToOne: false
            referencedRelation: "store_services"
            referencedColumns: ["service_id"]
          },
          {
            foreignKeyName: "repair_order_line_store_id_fkey"
            columns: ["store_id"]
            isOneToOne: false
            referencedRelation: "store"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "repair_order_line_store_id_fkey"
            columns: ["store_id"]
            isOneToOne: false
            referencedRelation: "store_services"
            referencedColumns: ["id"]
          },
        ]
      }
      repair_order_part: {
        Row: {
          cost: number | null
          created_at: string
          id: number
          name: string
          part_id: number | null
          part_no: string | null
          price: number
          quantity: number
          repair_order_id: string
          repair_order_line_id: number | null
          store_id: string
          unit_measure: string | null
          updated_at: string | null
        }
        Insert: {
          cost?: number | null
          created_at?: string
          id?: number
          name: string
          part_id?: number | null
          part_no?: string | null
          price?: number
          quantity?: number
          repair_order_id: string
          repair_order_line_id?: number | null
          store_id: string
          unit_measure?: string | null
          updated_at?: string | null
        }
        Update: {
          cost?: number | null
          created_at?: string
          id?: number
          name?: string
          part_id?: number | null
          part_no?: string | null
          price?: number
          quantity?: number
          repair_order_id?: string
          repair_order_line_id?: number | null
          store_id?: string
          unit_measure?: string | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "repair_order_line_store_id_fkey"
            columns: ["store_id"]
            isOneToOne: false
            referencedRelation: "store"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "repair_order_line_store_id_fkey"
            columns: ["store_id"]
            isOneToOne: false
            referencedRelation: "store_services"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "repair_order_part_part_id_fkey"
            columns: ["part_id"]
            isOneToOne: false
            referencedRelation: "parts"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "repair_order_part_repair_order_id_fkey"
            columns: ["repair_order_id"]
            isOneToOne: false
            referencedRelation: "repair_order"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "repair_order_part_repair_order_id_fkey"
            columns: ["repair_order_id"]
            isOneToOne: false
            referencedRelation: "repair_orders"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "repair_order_part_repair_order_line_id_fkey"
            columns: ["repair_order_line_id"]
            isOneToOne: false
            referencedRelation: "repair_order_line"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      notification_details: {
        Row: {
          category_name: string | null
          categoryId: number | null
          notification_id: number | null
          subCategoryId: number | null
          title: string | null
        }
        Relationships: []
      }
      repair_orders: {
        Row: {
          appointment_date: string | null
          appointment_date_str: string | null
          appointment_time: string | null
          brand_img_url: string | null
          created_at: string | null
          first_name: string | null
          id: string | null
          invoice_status: Database["public"]["Enums"]["invoice_status"] | null
          last_name: string | null
          make: string | null
          model: string | null
          order_line_parts: Json[] | null
          order_line_services: Json[] | null
          parts_total: number | null
          phone: string | null
          reference_no: number | null
          services_total: number | null
          status: Database["public"]["Enums"]["repair_order_status"] | null
          store_id: string | null
          total_cost: number | null
          user_id: string | null
          vehicle_id: string | null
          year_model: string | null
        }
        Relationships: [
          {
            foreignKeyName: "repair_order_store_id_fkey"
            columns: ["store_id"]
            isOneToOne: false
            referencedRelation: "store"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "repair_order_store_id_fkey"
            columns: ["store_id"]
            isOneToOne: false
            referencedRelation: "store_services"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "repair_order_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "repair_order_vehicle_id_fkey"
            columns: ["vehicle_id"]
            isOneToOne: false
            referencedRelation: "vehicle"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never